Refactoring admin controllers and admin.php routes file according to it's operations

+++ app/Http/Requests/TicketRequest.php
@@ -23,7 +23,7 @@ class TicketRequest extends FormRequest
     {
         return [
             // To be removed
-            'amount' => ['required', 'double']
+            'amount' => ['required', 'numeric']
             
             // PAYMENT DETAILS
         ];
